stress-gpu: ensure usec computations are using uint64_t type
Fix uint32_t integer overflow on multiply

+++ stress-gpu.c
@@ -523,8 +523,8 @@ static void stress_gpu_alarm_handler(int sig)
 static void *stress_gpu_pthread(void *arg)
 {
 	stress_args_t *args = (stress_args_t *)arg;
-	uint64_t sleep_usecs = 100000 * args->num_instances;
-	uint64_t start_sleep_usecs = 100000 * args->instance;
+	uint64_t sleep_usecs = 100000UL * (uint64_t)args->num_instances;
+	uint64_t start_sleep_usecs = 100000UL * (uint64_t)args->instance;
 
 	shim_usleep(start_sleep_usecs);
 	while (stress_continue(args)) {
